Refactor FPC metrics logging.
