+++ plugins/remotelogmetrics/plugin.go
@@ -4,6 +4,7 @@
 package remotelogmetrics
 
 import (
+	"github.com/iotaledger/goshimmer/plugins/messagelayer"
 	"sync"
 	"time"
 
@@ -36,14 +37,8 @@ func Plugin() *node.Plugin {
 }
 
 func configure(_ *node.Plugin) {
-	remotelogmetrics.Events().SyncBeaconSyncChanged.Attach(events.NewClosure(func(syncUpdate remotelogmetrics.SyncStatusChangedEvent) {
-		isSyncBeaconSynced.Store(syncUpdate.CurrentStatus)
-	}))
-	remotelogmetrics.Events().TangleTimeSyncChanged.Attach(events.NewClosure(func(syncUpdate remotelogmetrics.SyncStatusChangedEvent) {
-		isTangleTimeSynced.Store(syncUpdate.CurrentStatus)
-	}))
-	remotelogmetrics.Events().TangleTimeSyncChanged.Attach(events.NewClosure(sendSyncStatusChangedEvent))
-	remotelogmetrics.Events().SyncBeaconSyncChanged.Attach(events.NewClosure(sendSyncStatusChangedEvent))
+	configureSyncMetrics()
+	configureFPCConflictsMetrics()
 }
 
 func run(_ *node.Plugin) {
@@ -62,9 +57,26 @@ func run(_ *node.Plugin) {
 	}
 }
 
+func configureSyncMetrics() {
+	remotelogmetrics.Events().SyncBeaconSyncChanged.Attach(events.NewClosure(func(syncUpdate remotelogmetrics.SyncStatusChangedEvent) {
+		isSyncBeaconSynced.Store(syncUpdate.CurrentStatus)
+	}))
+	remotelogmetrics.Events().TangleTimeSyncChanged.Attach(events.NewClosure(func(syncUpdate remotelogmetrics.SyncStatusChangedEvent) {
+		isTangleTimeSynced.Store(syncUpdate.CurrentStatus)
+	}))
+	remotelogmetrics.Events().TangleTimeSyncChanged.Attach(events.NewClosure(sendSyncStatusChangedEvent))
+	remotelogmetrics.Events().SyncBeaconSyncChanged.Attach(events.NewClosure(sendSyncStatusChangedEvent))
+}
+
 func sendSyncStatusChangedEvent(syncUpdate remotelogmetrics.SyncStatusChangedEvent) {
 	err := remotelog.RemoteLogger().Send(syncUpdate)
 	if err != nil {
 		plugin.Logger().Errorw("Failed to send sync status changed record on sync change event.", "err", err)
 	}
+}
+
+func configureFPCConflictsMetrics() {
+	metricsLogger := newFPCMetricsLogger()
+	messagelayer.Voter().Events().Finalized.Attach(events.NewClosure(metricsLogger.onVoteFinalized))
+	messagelayer.Voter().Events().RoundExecuted.Attach(events.NewClosure(metricsLogger.onVoteRoundExecuted))
 }
